Factoring out ExtSearchController logic and introducing com.keepit.search.augmentation package
