+++ kifi-backend/modules/common/app/com/keepit/curator/model/SeedAttribution.scala
@@ -2,7 +2,6 @@ package com.keepit.curator.model
 
 import com.keepit.model.{ Keep, User, Library }
 import com.keepit.common.db.Id
-import com.keepit.search.{ RestrictedKeepInfo, AugmentationInfo }
 
 import com.kifi.macros.json
