+++ kifi-backend/modules/common/app/com/keepit/search/SearchServiceClient.scala
@@ -21,6 +21,7 @@ import com.keepit.social.{ BasicUser, TypeaheadUserHit }
 import com.keepit.typeahead.PrefixMatching
 import com.keepit.typeahead.PrefixFilter
 import scala.collection.mutable.ListBuffer
+import com.keepit.search.augmentation.{ ItemAugmentationResponse, ItemAugmentationRequest, AugmentableItem }
 
 trait SearchServiceClient extends ServiceClient {
   final val serviceType = ServiceType.SEARCH
