+++ kifi-backend/modules/search/app/com/keepit/controllers/ext/ExtSearchController.scala
@@ -9,27 +9,18 @@ import com.keepit.controllers.util.SearchControllerUtil
 import com.keepit.model._
 import com.keepit.model.ExperimentType.ADMIN
 import com.keepit.search.graph.library.LibraryIndexer
-import com.keepit.search.{ AugmentedItem, AugmentationCommander, SearchCommander }
+import com.keepit.search.{ SearchCommander }
 import com.keepit.shoebox.ShoeboxServiceClient
 import play.api.libs.concurrent.Execution.Implicits.defaultContext
 import play.api.libs.iteratee.Enumerator
 import play.api.libs.json._
-import play.api.libs.json.JsBoolean
-import play.api.libs.json.JsObject
-import com.keepit.common.db.Id
+import com.keepit.search.augmentation.{ AugmentedItem, AugmentationCommander }
+import com.keepit.common.json
 
 object ExtSearchController {
   private[ExtSearchController] val maxKeepersShown = 20
   private[ExtSearchController] val maxLibrariesShown = 10
   private[ExtSearchController] val maxTagsShown = 15
-
-  @inline private def canBeOmitted(value: JsValue): Boolean = value match {
-    case JsNull | JsBoolean(false) | JsString("") | JsArray(Seq()) => true
-    case JsNumber(zero) if zero == 0 => true
-    case _ => false
-  }
-
-  def compactJson(fullJson: JsObject): JsObject = JsObject(fullJson.fields.filterNot { case (key, value) => canBeOmitted(value) })
 }
 
 class ExtSearchController @Inject() (
@@ -87,51 +78,22 @@ class ExtSearchController @Inject() (
 
     val augmentationFuture = plainResultFuture.flatMap { kifiPlainResult =>
       augment(augmentationCommander, userId, kifiPlainResult).flatMap { augmentedItems =>
-        val allKeepersShown = augmentedItems.map(_.relatedKeepers.take(maxKeepersShown))
-        val allLibrariesShown = augmentedItems.map(_.libraries.take(maxLibrariesShown))
-
-        val userIds = ((allKeepersShown.flatten ++ allLibrariesShown.flatMap(_.map(_._2))).toSet - userId).toSeq
-        val futureUsers = shoeboxClient.getBasicUsers(userIds)
-        val userIndexById = userIds.zipWithIndex.toMap + (userId -> -1)
-
-        val libraryIds = allLibrariesShown.flatMap(_.map(_._1)).distinct
-        val libraryIndexById = libraryIds.zipWithIndex.toMap
         val librarySearcher = libraryIndexer.getSearcher
+
+        val (allSecondaryFields, userIds, libraryIds) = AugmentedItem.writesAugmentationFields(librarySearcher, userId, maxKeepersShown, maxLibrariesShown, maxTagsShown, augmentedItems)
+
+        val futureUsers = shoeboxClient.getBasicUsers(userIds)
+
+        val hitsJson = allSecondaryFields.map(json.minify)
+
         val libraries = {
           val libraryById = getBasicLibraries(librarySearcher, libraryIds.toSet)
           libraryIds.map(libraryById(_))
         }
 
-        val hitsJson = augmentedItems.zipWithIndex.map {
-          case (augmentedItem, itemIndex) =>
-            val secret = augmentedItem.isSecret(librarySearcher)
-
-            val keepersShown = allKeepersShown(itemIndex).map(userIndexById(_))
-            val keepersOmitted = augmentedItem.relatedKeepers.size - keepersShown.size
-            val keepersTotal = augmentedItem.keepersTotal
-
-            val librariesShown = allLibrariesShown(itemIndex).map { case (libraryId, keeperId) => (libraryIndexById(libraryId), userIndexById(keeperId)) }
-            val librariesOmitted = augmentedItem.libraries.size - librariesShown.size
-
-            val tagsShown = augmentedItem.tags.take(maxTagsShown)
-            val tagsOmitted = augmentedItem.tags.size - tagsShown.size
-
-            val fullJson = Json.obj(
-              "secret" -> JsBoolean(secret),
-              "keepers" -> keepersShown,
-              "keepersOmitted" -> keepersOmitted,
-              "keepersTotal" -> keepersTotal,
-              "libraries" -> librariesShown.flatMap { case (libraryIndex, keeperIndex) => Seq(libraryIndex, keeperIndex) },
-              "librariesOmitted" -> librariesOmitted,
-              "tags" -> tagsShown,
-              "tagsOmitted" -> tagsOmitted
-            )
-            compactJson(fullJson)
-        }
         futureUsers.map { usersById =>
           val users = userIds.map(usersById(_))
-          val librariesJson = libraries.map { library => compactJson(Json.obj("id" -> library.id, "name" -> library.name, "secret" -> library.isSecret)) }
-          Json.obj("hits" -> hitsJson, "libraries" -> librariesJson, "users" -> users)
+          Json.obj("hits" -> hitsJson, "users" -> users, "libraries" -> libraries)
         }
       }
     }
